+++ engine/src/communication/CommunicationInterface/messageReceiver.hpp
@@ -11,7 +11,6 @@
 #include "execution_graph/logic_controllers/CacheMachine.h"
 
 
-
 namespace comm {
 
 /**