+++ engine/src/execution_graph/logic_controllers/taskflow/executor.cpp
@@ -1,4 +1,5 @@
 #include "executor.h"
+#include "execution_graph/logic_controllers/GPUCacheData.h"
 
 using namespace fmt::literals;